+++ tests/Unit/Private/Resolve-Reason.Tests.ps1
@@ -42,11 +42,11 @@ AfterAll {
     Get-Module -Name $script:dscModuleName -All | Remove-Module -Force
 }
 
-Describe 'New-Reason' -Tag 'Private' {
+Describe 'Resolve-Reason' -Tag 'Private' {
     Context 'When passing an empty collection' {
         It 'Should return an empty collection' {
             InModuleScope -ScriptBlock {
-                $result = New-Reason -Property @() -ResourceName 'MyResource'
+                $result = Resolve-Reason -Property @() -ResourceName 'MyResource'
 
                 $result | Should -HaveCount 0
             }
@@ -56,7 +56,7 @@ Describe 'New-Reason' -Tag 'Private' {
     Context 'When passing a null value' {
         It 'Should return an empty collection' {
             InModuleScope -ScriptBlock {
-                $result = New-Reason -Property $null -ResourceName 'MyResource'
+                $result = Resolve-Reason -Property $null -ResourceName 'MyResource'
 
                 $result | Should -HaveCount 0
             }
@@ -79,7 +79,7 @@ Describe 'New-Reason' -Tag 'Private' {
                     }
                 )
 
-                $result = New-Reason -Property $mockProperties -ResourceName 'MyResource'
+                $result = Resolve-Reason -Property $mockProperties -ResourceName 'MyResource'
 
                 $result | Should -HaveCount 2
 
@@ -108,7 +108,7 @@ Describe 'New-Reason' -Tag 'Private' {
                     }
                 )
 
-                $result = $mockProperties | New-Reason -ResourceName 'MyResource'
+                $result = $mockProperties | Resolve-Reason -ResourceName 'MyResource'
 
                 $result | Should -HaveCount 2
 
@@ -142,7 +142,7 @@ Describe 'New-Reason' -Tag 'Private' {
                         }
                     )
 
-                    $result = New-Reason -Property $mockProperties -ResourceName 'MyResource'
+                    $result = Resolve-Reason -Property $mockProperties -ResourceName 'MyResource'
 
                     $result | Should -HaveCount 1
 
@@ -175,7 +175,7 @@ Describe 'New-Reason' -Tag 'Private' {
                         }
                     )
 
-                    $result = New-Reason -Property $mockProperties -ResourceName 'MyResource'
+                    $result = Resolve-Reason -Property $mockProperties -ResourceName 'MyResource'
 
                     $result | Should -HaveCount 1
 
@@ -197,7 +197,7 @@ Describe 'New-Reason' -Tag 'Private' {
                     }
                 )
 
-                $result = New-Reason -Property $mockProperties -ResourceName 'MyResource'
+                $result = Resolve-Reason -Property $mockProperties -ResourceName 'MyResource'
 
                 $result | Should -HaveCount 1